storage: move processRaftCommand
Release note: None
