+++ pkg/storage/replica_raft.go
@@ -1441,6 +1441,232 @@ func (r *Replica) checkForcedErrLocked(
 	return leaseIndex, proposalNoReevaluation, nil
 }
 
+type snapTruncationInfo struct {
+	index    uint64
+	deadline time.Time
+}
+
+func (r *Replica) addSnapshotLogTruncationConstraintLocked(
+	ctx context.Context, snapUUID uuid.UUID, index uint64,
+) {
+	if r.mu.snapshotLogTruncationConstraints == nil {
+		r.mu.snapshotLogTruncationConstraints = make(map[uuid.UUID]snapTruncationInfo)
+	}
+	item, ok := r.mu.snapshotLogTruncationConstraints[snapUUID]
+	if ok {
+		// Uh-oh, there's either a programming error (resulting in the same snapshot
+		// fed into this method twice) or a UUID collision. We discard the update
+		// (which is benign) but log it loudly. If the index is the same, it's
+		// likely the former, otherwise the latter.
+		log.Warningf(ctx, "UUID collision at %s for %+v (index %d)", snapUUID, item, index)
+		return
+	}
+
+	r.mu.snapshotLogTruncationConstraints[snapUUID] = snapTruncationInfo{index: index}
+}
+
+func (r *Replica) completeSnapshotLogTruncationConstraint(
+	ctx context.Context, snapUUID uuid.UUID, now time.Time,
+) {
+	deadline := now.Add(raftLogQueuePendingSnapshotGracePeriod)
+
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	item, ok := r.mu.snapshotLogTruncationConstraints[snapUUID]
+	if !ok {
+		// UUID collision while adding the snapshot in originally. Nothing
+		// else to do.
+		return
+	}
+
+	item.deadline = deadline
+	r.mu.snapshotLogTruncationConstraints[snapUUID] = item
+}
+
+func (r *Replica) getAndGCSnapshotLogTruncationConstraintsLocked(
+	now time.Time,
+) (minSnapIndex uint64) {
+	for snapUUID, item := range r.mu.snapshotLogTruncationConstraints {
+		if item.deadline != (time.Time{}) && item.deadline.Before(now) {
+			// The snapshot has finished and its grace period has passed.
+			// Ignore it when making truncation decisions.
+			delete(r.mu.snapshotLogTruncationConstraints, snapUUID)
+			continue
+		}
+		if minSnapIndex == 0 || minSnapIndex > item.index {
+			minSnapIndex = item.index
+		}
+	}
+	if len(r.mu.snapshotLogTruncationConstraints) == 0 {
+		// Save a little bit of memory.
+		r.mu.snapshotLogTruncationConstraints = nil
+	}
+	return minSnapIndex
+}
+
+func isRaftLeader(raftStatus *raft.Status) bool {
+	return raftStatus != nil && raftStatus.SoftState.RaftState == raft.StateLeader
+}
+
+// HasRaftLeader returns true if the raft group has a raft leader currently.
+func HasRaftLeader(raftStatus *raft.Status) bool {
+	return raftStatus != nil && raftStatus.SoftState.Lead != 0
+}
+
+// pendingCmdSlice sorts by increasing MaxLeaseIndex.
+type pendingCmdSlice []*ProposalData
+
+func (s pendingCmdSlice) Len() int      { return len(s) }
+func (s pendingCmdSlice) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
+func (s pendingCmdSlice) Less(i, j int) bool {
+	return s[i].command.MaxLeaseIndex < s[j].command.MaxLeaseIndex
+}
+
+// withRaftGroupLocked calls the supplied function with the (lazily
+// initialized) Raft group. The supplied function should return true for the
+// unquiesceAndWakeLeader argument if the replica should be unquiesced (and the
+// leader awoken). See handleRaftReady for an instance of where this value
+// varies.
+//
+// Requires that Replica.mu is held. Also requires that Replica.raftMu is held
+// if either the caller can't guarantee that r.mu.internalRaftGroup != nil or
+// the provided function requires Replica.raftMu.
+func (r *Replica) withRaftGroupLocked(
+	mayCampaignOnWake bool, f func(r *raft.RawNode) (unquiesceAndWakeLeader bool, _ error),
+) error {
+	if r.mu.destroyStatus.RemovedOrCorrupt() {
+		// Silently ignore all operations on destroyed replicas. We can't return an
+		// error here as all errors returned from this method are considered fatal.
+		return nil
+	}
+
+	if r.mu.replicaID == 0 {
+		// The replica's raft group has not yet been configured (i.e. the replica
+		// was created from a preemptive snapshot).
+		return nil
+	}
+
+	if r.mu.internalRaftGroup == nil {
+		r.raftMu.Mutex.AssertHeld()
+
+		ctx := r.AnnotateCtx(context.TODO())
+		raftGroup, err := raft.NewRawNode(newRaftConfig(
+			raft.Storage((*replicaRaftStorage)(r)),
+			uint64(r.mu.replicaID),
+			r.mu.state.RaftAppliedIndex,
+			r.store.cfg,
+			&raftLogger{ctx: ctx},
+		), nil)
+		if err != nil {
+			return err
+		}
+		r.mu.internalRaftGroup = raftGroup
+
+		if mayCampaignOnWake {
+			r.maybeCampaignOnWakeLocked(ctx)
+		}
+	}
+
+	// This wrapper function is a hack to add range IDs to stack traces
+	// using the same pattern as Replica.sendWithRangeID.
+	unquiesce, err := func(rangeID roachpb.RangeID, raftGroup *raft.RawNode) (bool, error) {
+		return f(raftGroup)
+	}(r.RangeID, r.mu.internalRaftGroup)
+	if unquiesce {
+		r.unquiesceAndWakeLeaderLocked()
+	}
+	return err
+}
+
+// withRaftGroup calls the supplied function with the (lazily initialized)
+// Raft group. It acquires and releases the Replica lock, so r.mu must not be
+// held (or acquired by the supplied function).
+//
+// If mayCampaignOnWake is true, the replica may initiate a raft
+// election if it was previously in a dormant state. Most callers
+// should set this to true, because the prevote feature minimizes the
+// disruption from unnecessary elections. The exception is that we
+// should not initiate an election while handling incoming raft
+// messages (which may include MsgVotes from an election in progress,
+// and this election would be disrupted if we started our own).
+//
+// Has the same requirement for Replica.raftMu as withRaftGroupLocked.
+func (r *Replica) withRaftGroup(
+	mayCampaignOnWake bool, f func(r *raft.RawNode) (unquiesceAndWakeLeader bool, _ error),
+) error {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	return r.withRaftGroupLocked(mayCampaignOnWake, f)
+}
+
+func shouldCampaignOnWake(
+	leaseStatus storagepb.LeaseStatus,
+	lease roachpb.Lease,
+	storeID roachpb.StoreID,
+	raftStatus raft.Status,
+) bool {
+	// When waking up a range, campaign unless we know that another
+	// node holds a valid lease (this is most important after a split,
+	// when all replicas create their raft groups at about the same
+	// time, with a lease pre-assigned to one of them). Note that
+	// thanks to PreVote, unnecessary campaigns are not disruptive so
+	// we should err on the side of campaigining here.
+	anotherOwnsLease := leaseStatus.State == storagepb.LeaseState_VALID && !lease.OwnedBy(storeID)
+
+	// If we're already campaigning or know who the leader is, don't
+	// start a new term.
+	noLeader := raftStatus.RaftState == raft.StateFollower && raftStatus.Lead == 0
+	return !anotherOwnsLease && noLeader
+}
+
+// maybeCampaignOnWakeLocked is called when the range wakes from a
+// dormant state (either the initial "raftGroup == nil" state or after
+// being quiescent) and campaigns for raft leadership if appropriate.
+func (r *Replica) maybeCampaignOnWakeLocked(ctx context.Context) {
+	// Raft panics if a node that is not currently a member of the
+	// group tries to campaign. That happens primarily when we apply
+	// preemptive snapshots.
+	if _, currentMember := r.mu.state.Desc.GetReplicaDescriptorByID(r.mu.replicaID); !currentMember {
+		return
+	}
+
+	leaseStatus := r.leaseStatus(*r.mu.state.Lease, r.store.Clock().Now(), r.mu.minLeaseProposedTS)
+	raftStatus := r.mu.internalRaftGroup.Status()
+	if shouldCampaignOnWake(leaseStatus, *r.mu.state.Lease, r.store.StoreID(), *raftStatus) {
+		log.VEventf(ctx, 3, "campaigning")
+		if err := r.mu.internalRaftGroup.Campaign(); err != nil {
+			log.VEventf(ctx, 1, "failed to campaign: %s", err)
+		}
+	}
+}
+
+// a lastUpdateTimesMap is maintained on the Raft leader to keep track of the
+// last communication received from followers, which in turn informs the quota
+// pool and log truncations.
+type lastUpdateTimesMap map[roachpb.ReplicaID]time.Time
+
+func (m lastUpdateTimesMap) update(replicaID roachpb.ReplicaID, now time.Time) {
+	if m == nil {
+		return
+	}
+	m[replicaID] = now
+}
+
+// isFollowerActive returns whether the specified follower has made
+// communication with the leader in the last MaxQuotaReplicaLivenessDuration.
+func (m lastUpdateTimesMap) isFollowerActive(
+	ctx context.Context, replicaID roachpb.ReplicaID, now time.Time,
+) bool {
+	lastUpdateTime, ok := m[replicaID]
+	if !ok {
+		// If the follower has no entry in lastUpdateTimes, it has not been
+		// updated since r became the leader (at which point all then-existing
+		// replicas were updated).
+		return false
+	}
+	return now.Sub(lastUpdateTime) <= MaxQuotaReplicaLivenessDuration
+}
+
 // processRaftCommand processes a raft command by unpacking the
 // command struct to get args and reply and then applying the command
 // to the state machine via applyRaftCommand(). The result is sent on
@@ -2089,85 +2315,4 @@ func (r *Replica) applyRaftCommand(
 	elapsed := timeutil.Since(start)
 	r.store.metrics.RaftCommandCommitLatency.RecordValue(elapsed.Nanoseconds())
 	return deltaStats, nil
-}
-
-type snapTruncationInfo struct {
-	index    uint64
-	deadline time.Time
-}
-
-func (r *Replica) addSnapshotLogTruncationConstraintLocked(
-	ctx context.Context, snapUUID uuid.UUID, index uint64,
-) {
-	if r.mu.snapshotLogTruncationConstraints == nil {
-		r.mu.snapshotLogTruncationConstraints = make(map[uuid.UUID]snapTruncationInfo)
-	}
-	item, ok := r.mu.snapshotLogTruncationConstraints[snapUUID]
-	if ok {
-		// Uh-oh, there's either a programming error (resulting in the same snapshot
-		// fed into this method twice) or a UUID collision. We discard the update
-		// (which is benign) but log it loudly. If the index is the same, it's
-		// likely the former, otherwise the latter.
-		log.Warningf(ctx, "UUID collision at %s for %+v (index %d)", snapUUID, item, index)
-		return
-	}
-
-	r.mu.snapshotLogTruncationConstraints[snapUUID] = snapTruncationInfo{index: index}
-}
-
-func (r *Replica) completeSnapshotLogTruncationConstraint(
-	ctx context.Context, snapUUID uuid.UUID, now time.Time,
-) {
-	deadline := now.Add(raftLogQueuePendingSnapshotGracePeriod)
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	item, ok := r.mu.snapshotLogTruncationConstraints[snapUUID]
-	if !ok {
-		// UUID collision while adding the snapshot in originally. Nothing
-		// else to do.
-		return
-	}
-
-	item.deadline = deadline
-	r.mu.snapshotLogTruncationConstraints[snapUUID] = item
-}
-
-func (r *Replica) getAndGCSnapshotLogTruncationConstraintsLocked(
-	now time.Time,
-) (minSnapIndex uint64) {
-	for snapUUID, item := range r.mu.snapshotLogTruncationConstraints {
-		if item.deadline != (time.Time{}) && item.deadline.Before(now) {
-			// The snapshot has finished and its grace period has passed.
-			// Ignore it when making truncation decisions.
-			delete(r.mu.snapshotLogTruncationConstraints, snapUUID)
-			continue
-		}
-		if minSnapIndex == 0 || minSnapIndex > item.index {
-			minSnapIndex = item.index
-		}
-	}
-	if len(r.mu.snapshotLogTruncationConstraints) == 0 {
-		// Save a little bit of memory.
-		r.mu.snapshotLogTruncationConstraints = nil
-	}
-	return minSnapIndex
-}
-
-func isRaftLeader(raftStatus *raft.Status) bool {
-	return raftStatus != nil && raftStatus.SoftState.RaftState == raft.StateLeader
-}
-
-// HasRaftLeader returns true if the raft group has a raft leader currently.
-func HasRaftLeader(raftStatus *raft.Status) bool {
-	return raftStatus != nil && raftStatus.SoftState.Lead != 0
-}
-
-// pendingCmdSlice sorts by increasing MaxLeaseIndex.
-type pendingCmdSlice []*ProposalData
-
-func (s pendingCmdSlice) Len() int      { return len(s) }
-func (s pendingCmdSlice) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
-func (s pendingCmdSlice) Less(i, j int) bool {
-	return s[i].command.MaxLeaseIndex < s[j].command.MaxLeaseIndex
 }
