+++ pkg/storage/replica_range_lease.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage/storagepb"
@@ -649,6 +650,323 @@ func (r *Replica) AdminTransferLease(ctx context.Context, target roachpb.StoreID
 		case <-ctx.Done():
 			extension.Cancel()
 			return ctx.Err()
+		}
+	}
+}
+
+// GetLease returns the lease and, if available, the proposed next lease.
+func (r *Replica) GetLease() (roachpb.Lease, roachpb.Lease) {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	return r.getLeaseRLocked()
+}
+
+func (r *Replica) getLeaseRLocked() (roachpb.Lease, roachpb.Lease) {
+	if nextLease, ok := r.mu.pendingLeaseRequest.RequestPending(); ok {
+		return *r.mu.state.Lease, nextLease
+	}
+	return *r.mu.state.Lease, roachpb.Lease{}
+}
+
+// OwnsValidLease returns whether this replica is the current valid
+// leaseholder. Note that this method does not check to see if a transfer is
+// pending, but returns the status of the current lease and ownership at the
+// specified point in time.
+func (r *Replica) OwnsValidLease(ts hlc.Timestamp) bool {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	return r.ownsValidLeaseRLocked(ts)
+}
+
+func (r *Replica) ownsValidLeaseRLocked(ts hlc.Timestamp) bool {
+	return r.mu.state.Lease.OwnedBy(r.store.StoreID()) &&
+		r.leaseStatus(*r.mu.state.Lease, ts, r.mu.minLeaseProposedTS).State == storagepb.LeaseState_VALID
+}
+
+// IsLeaseValid returns true if the replica's lease is owned by this
+// replica and is valid (not expired, not in stasis).
+func (r *Replica) IsLeaseValid(lease roachpb.Lease, ts hlc.Timestamp) bool {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	return r.isLeaseValidRLocked(lease, ts)
+}
+
+func (r *Replica) isLeaseValidRLocked(lease roachpb.Lease, ts hlc.Timestamp) bool {
+	return r.leaseStatus(lease, ts, r.mu.minLeaseProposedTS).State == storagepb.LeaseState_VALID
+}
+
+// newNotLeaseHolderError returns a NotLeaseHolderError initialized with the
+// replica for the holder (if any) of the given lease.
+//
+// Note that this error can be generated on the Raft processing goroutine, so
+// its output should be completely determined by its parameters.
+func newNotLeaseHolderError(
+	l *roachpb.Lease, proposerStoreID roachpb.StoreID, rangeDesc *roachpb.RangeDescriptor,
+) *roachpb.NotLeaseHolderError {
+	err := &roachpb.NotLeaseHolderError{
+		RangeID: rangeDesc.RangeID,
+	}
+	err.Replica, _ = rangeDesc.GetReplicaDescriptor(proposerStoreID)
+	if l != nil {
+		// Normally, we return the lease-holding Replica here. However, in the
+		// case in which a leader removes itself, we want the followers to
+		// avoid handing out a misleading clue (which in itself shouldn't be
+		// overly disruptive as the lease would expire and then this method
+		// shouldn't be called for it any more, but at the very least it
+		// could catch tests in a loop, presumably due to manual clocks).
+		_, stillMember := rangeDesc.GetReplicaDescriptor(l.Replica.StoreID)
+		if stillMember {
+			err.LeaseHolder = &l.Replica
+			err.Lease = l
+		}
+	}
+	return err
+}
+
+// leaseGoodToGo is a fast-path for lease checks which verifies that an
+// existing lease is valid and owned by the current store. This method should
+// not be called directly. Use redirectOnOrAcquireLease instead.
+func (r *Replica) leaseGoodToGo(ctx context.Context) (storagepb.LeaseStatus, bool) {
+	timestamp := r.store.Clock().Now()
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+
+	if r.requiresExpiringLeaseRLocked() {
+		// Slow-path for expiration-based leases.
+		return storagepb.LeaseStatus{}, false
+	}
+
+	status := r.leaseStatus(*r.mu.state.Lease, timestamp, r.mu.minLeaseProposedTS)
+	if status.State == storagepb.LeaseState_VALID && status.Lease.OwnedBy(r.store.StoreID()) {
+		// We own the lease...
+		if repDesc, err := r.getReplicaDescriptorRLocked(); err == nil {
+			if _, ok := r.mu.pendingLeaseRequest.TransferInProgress(repDesc.ReplicaID); !ok {
+				// ...and there is no transfer pending.
+				return status, true
+			}
+		}
+	}
+	return storagepb.LeaseStatus{}, false
+}
+
+// redirectOnOrAcquireLease checks whether this replica has the lease
+// at the current timestamp. If it does, returns success. If another
+// replica currently holds the lease, redirects by returning
+// NotLeaseHolderError. If the lease is expired, a renewal is
+// synchronously requested. Leases are eagerly renewed when a request
+// with a timestamp within rangeLeaseRenewalDuration of the lease
+// expiration is served.
+//
+// TODO(spencer): for write commands, don't wait while requesting
+//  the range lease. If the lease acquisition fails, the write cmd
+//  will fail as well. If it succeeds, as is likely, then the write
+//  will not incur latency waiting for the command to complete.
+//  Reads, however, must wait.
+func (r *Replica) redirectOnOrAcquireLease(
+	ctx context.Context,
+) (storagepb.LeaseStatus, *roachpb.Error) {
+	if status, ok := r.leaseGoodToGo(ctx); ok {
+		return status, nil
+	}
+
+	// Loop until the lease is held or the replica ascertains the actual
+	// lease holder. Returns also on context.Done() (timeout or cancellation).
+	var status storagepb.LeaseStatus
+	for attempt := 1; ; attempt++ {
+		timestamp := r.store.Clock().Now()
+		llHandle, pErr := func() (*leaseRequestHandle, *roachpb.Error) {
+			r.mu.Lock()
+			defer r.mu.Unlock()
+
+			status = r.leaseStatus(*r.mu.state.Lease, timestamp, r.mu.minLeaseProposedTS)
+			switch status.State {
+			case storagepb.LeaseState_ERROR:
+				// Lease state couldn't be determined.
+				log.VEventf(ctx, 2, "lease state couldn't be determined")
+				return nil, roachpb.NewError(
+					newNotLeaseHolderError(nil, r.store.StoreID(), r.mu.state.Desc))
+
+			case storagepb.LeaseState_VALID, storagepb.LeaseState_STASIS:
+				if !status.Lease.OwnedBy(r.store.StoreID()) {
+					_, stillMember := r.mu.state.Desc.GetReplicaDescriptor(status.Lease.Replica.StoreID)
+					if !stillMember {
+						// This would be the situation in which the lease holder gets removed when
+						// holding the lease, or in which a lease request erroneously gets accepted
+						// for a replica that is not in the replica set. Neither of the two can
+						// happen in normal usage since appropriate mechanisms have been added:
+						//
+						// 1. Only the lease holder (at the time) schedules removal of a replica,
+						// but the lease can change hands and so the situation in which a follower
+						// coordinates a replica removal of the (new) lease holder is possible (if
+						// unlikely) in practice. In this situation, the new lease holder would at
+						// some point be asked to propose the replica change's EndTransaction to
+						// Raft. A check has been added that prevents proposals that amount to the
+						// removal of the proposer's (and hence lease holder's) Replica, preventing
+						// this scenario.
+						//
+						// 2. A lease is accepted for a Replica that has been removed. Without
+						// precautions, this could happen because lease requests are special in
+						// that they are the only command that is proposed on a follower (other
+						// commands may be proposed from followers, but not successfully so). For
+						// all proposals, processRaftCommand checks that their ProposalLease is
+						// compatible with the active lease for the log position. For commands
+						// proposed on the lease holder, the spanlatch manager then serializes
+						// everything. But lease requests get created on followers based on their
+						// local state and thus without being sequenced through latching. Thus
+						// a recently removed follower (unaware of its own removal) could submit
+						// a proposal for the lease (correctly using as a ProposerLease the last
+						// active lease), and would receive it given the up-to-date ProposerLease.
+						// Hence, an extra check is in order: processRaftCommand makes sure that
+						// lease requests for a replica not in the descriptor are bounced.
+						//
+						// However, this is possible if the `cockroach debug
+						// unsafe-remove-dead-replicas` command has been used, so
+						// this is just a logged error instead of a fatal
+						// assertion.
+						log.Errorf(ctx, "lease %s owned by replica %+v that no longer exists",
+							status.Lease, status.Lease.Replica)
+					}
+					// Otherwise, if the lease is currently held by another replica, redirect
+					// to the holder.
+					return nil, roachpb.NewError(
+						newNotLeaseHolderError(&status.Lease, r.store.StoreID(), r.mu.state.Desc))
+				}
+				// Check that we're not in the process of transferring the lease away.
+				// If we are transferring the lease away, we can't serve reads or
+				// propose Raft commands - see comments on TransferLease.
+				// TODO(andrei): If the lease is being transferred, consider returning a
+				// new error type so the client backs off until the transfer is
+				// completed.
+				repDesc, err := r.getReplicaDescriptorRLocked()
+				if err != nil {
+					return nil, roachpb.NewError(err)
+				}
+				if transferLease, ok := r.mu.pendingLeaseRequest.TransferInProgress(
+					repDesc.ReplicaID); ok {
+					return nil, roachpb.NewError(
+						newNotLeaseHolderError(&transferLease, r.store.StoreID(), r.mu.state.Desc))
+				}
+
+				// If the lease is in stasis, we can't serve requests until we've
+				// renewed the lease, so we return the handle to block on renewal.
+				// Otherwise, we don't need to wait for the extension and simply
+				// ignore the returned handle (whose channel is buffered) and continue.
+				if status.State == storagepb.LeaseState_STASIS {
+					return r.requestLeaseLocked(ctx, status), nil
+				}
+
+				// Extend the lease if this range uses expiration-based
+				// leases, the lease is in need of renewal, and there's not
+				// already an extension pending.
+				_, requestPending := r.mu.pendingLeaseRequest.RequestPending()
+				if !requestPending && r.requiresExpiringLeaseRLocked() {
+					renewal := status.Lease.Expiration.Add(-r.store.cfg.RangeLeaseRenewalDuration().Nanoseconds(), 0)
+					if !timestamp.Less(renewal) {
+						if log.V(2) {
+							log.Infof(ctx, "extending lease %s at %s", status.Lease, timestamp)
+						}
+						// We had an active lease to begin with, but we want to trigger
+						// a lease extension. We explicitly ignore the returned handle
+						// as we won't block on it.
+						_ = r.requestLeaseLocked(ctx, status)
+					}
+				}
+
+			case storagepb.LeaseState_EXPIRED:
+				// No active lease: Request renewal if a renewal is not already pending.
+				log.VEventf(ctx, 2, "request range lease (attempt #%d)", attempt)
+				return r.requestLeaseLocked(ctx, status), nil
+
+			case storagepb.LeaseState_PROSCRIBED:
+				// Lease proposed timestamp is earlier than the min proposed
+				// timestamp limit this replica must observe. If this store
+				// owns the lease, re-request. Otherwise, redirect.
+				if status.Lease.OwnedBy(r.store.StoreID()) {
+					log.VEventf(ctx, 2, "request range lease (attempt #%d)", attempt)
+					return r.requestLeaseLocked(ctx, status), nil
+				}
+				// If lease is currently held by another, redirect to holder.
+				return nil, roachpb.NewError(
+					newNotLeaseHolderError(&status.Lease, r.store.StoreID(), r.mu.state.Desc))
+			}
+
+			// Return a nil handle to signal that we have a valid lease.
+			return nil, nil
+		}()
+		if pErr != nil {
+			return storagepb.LeaseStatus{}, pErr
+		}
+		if llHandle == nil {
+			// We own a valid lease.
+			return status, nil
+		}
+
+		// Wait for the range lease to finish, or the context to expire.
+		pErr = func() (pErr *roachpb.Error) {
+			slowTimer := timeutil.NewTimer()
+			defer slowTimer.Stop()
+			slowTimer.Reset(base.SlowRequestThreshold)
+			tBegin := timeutil.Now()
+			for {
+				select {
+				case pErr = <-llHandle.C():
+					if pErr != nil {
+						switch tErr := pErr.GetDetail().(type) {
+						case *roachpb.AmbiguousResultError:
+							// This can happen if the RequestLease command we sent has been
+							// applied locally through a snapshot: the RequestLeaseRequest
+							// cannot be reproposed so we get this ambiguity.
+							// We'll just loop around.
+							return nil
+						case *roachpb.LeaseRejectedError:
+							if tErr.Existing.OwnedBy(r.store.StoreID()) {
+								// The RequestLease command we sent was rejected because another
+								// lease was applied in the meantime, but we own that other
+								// lease. So, loop until the current node becomes aware that
+								// it's the leaseholder.
+								return nil
+							}
+
+							// Getting a LeaseRejectedError back means someone else got there
+							// first, or the lease request was somehow invalid due to a concurrent
+							// change. That concurrent change could have been that this replica was
+							// removed (see processRaftCommand), so check for that case before
+							// falling back to a NotLeaseHolderError.
+							var err error
+							if _, descErr := r.GetReplicaDescriptor(); descErr != nil {
+								err = descErr
+							} else if lease, _ := r.GetLease(); !r.IsLeaseValid(lease, r.store.Clock().Now()) {
+								err = newNotLeaseHolderError(nil, r.store.StoreID(), r.Desc())
+							} else {
+								err = newNotLeaseHolderError(&lease, r.store.StoreID(), r.Desc())
+							}
+							pErr = roachpb.NewError(err)
+						}
+						return pErr
+					}
+					log.Eventf(ctx, "lease acquisition succeeded: %+v", status.Lease)
+					return nil
+				case <-slowTimer.C:
+					slowTimer.Read = true
+					log.Warningf(ctx, "have been waiting %s attempting to acquire lease",
+						base.SlowRequestThreshold)
+					r.store.metrics.SlowLeaseRequests.Inc(1)
+					defer func() {
+						r.store.metrics.SlowLeaseRequests.Dec(1)
+						log.Infof(ctx, "slow lease acquisition finished after %s with error %v after %d attempts", timeutil.Since(tBegin), pErr, attempt)
+					}()
+				case <-ctx.Done():
+					llHandle.Cancel()
+					log.VErrEventf(ctx, 2, "lease acquisition failed: %s", ctx.Err())
+					return roachpb.NewError(newNotLeaseHolderError(nil, r.store.StoreID(), r.Desc()))
+				case <-r.store.Stopper().ShouldStop():
+					llHandle.Cancel()
+					return roachpb.NewError(newNotLeaseHolderError(nil, r.store.StoreID(), r.Desc()))
+				}
+			}
+		}()
+		if pErr != nil {
+			return storagepb.LeaseStatus{}, pErr
 		}
 	}
 }
